Add migrations for old production schema modification

+++ database/migrations/1763762387514_create_add_is_authorizer_to_users_table.ts
@@ -4,14 +4,20 @@ export default class extends BaseSchema {
   protected tableName = 'users'
 
   async up() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.boolean('is_authorizer').defaultTo(false)
-    })
+    const hasIsAuthorizer = await this.schema.hasColumn(this.tableName, 'is_authorizer')
+    if (!hasIsAuthorizer) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.boolean('is_authorizer').defaultTo(false)
+      })
+    }
   }
 
   async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn('is_authorizer')
-    })
+    const hasIsAuthorizer = await this.schema.hasColumn(this.tableName, 'is_authorizer')
+    if (hasIsAuthorizer) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropColumn('is_authorizer')
+      })
+    }
   }
 }
